requirements.txt
grpcio==1.46.0
grpcio-tools==1.46.0
protobuf==3.20.1
PyYAML==6.0
six==1.16.0
